+++ src/MinFraud/Util.php
@@ -267,6 +267,8 @@ class Util
             }
         }
 
+        $domain = preg_replace('/(?:\.com){2,}$/', '.com', $domain);
+
         if (isset(self::$typoDomains[$domain])) {
             $domain = self::$typoDomains[$domain];
         }